move reportversion classes into own file
